NGWMN-2099 -- Review changes:   Moved BaseController out of the Collections package.   Refactored BaseController determineContentType() to return the content type.   Format update: GlobalDefaultExceptionHandler   Order WaterML by: monitoring_location_identifier, observed_property_name, statistic
