+++ src/main/java/gov/usgs/wma/waterdata/BaseController.java
@@ -1,4 +1,4 @@
-package gov.usgs.wma.waterdata.collections;
+package gov.usgs.wma.waterdata;
 
 import gov.usgs.wma.waterdata.parameter.ContentType;
 
@@ -10,7 +10,6 @@ import org.springframework.web.HttpMediaTypeNotAcceptableException;
 public abstract class BaseController {
 	protected final String CONTENT_TYPE_MESS = "Content type (f=) must be one of: ";
 
-	protected ContentType contentType;
 	protected final String contentTypeDesc = "Content format returned, default json";
 	protected String ogc404Payload = "{\"code\":\"404\", \"description\":\"Requested data not found\"}";
 
@@ -18,35 +17,26 @@ public abstract class BaseController {
 	protected static final String HTTP_404_DESCRIPTION = "The requested data was not found.";
 	protected static final String HTTP_500_DESCRIPTION = "Unexpected error occurred.";
 
-	protected void determineContentType(String mimeType) throws HttpMediaTypeNotAcceptableException {
-		determineContentType(mimeType, Arrays.asList(ContentType.values()));
+	protected ContentType determineContentType(String mimeType) throws HttpMediaTypeNotAcceptableException {
+		return determineContentType(mimeType, Arrays.asList(ContentType.values()));
 	}
 
-	protected void determineContentType(String mimeType, List<ContentType> acceptedTypes)
+	protected ContentType determineContentType(String mimeType, List<ContentType> acceptedTypes)
 			throws HttpMediaTypeNotAcceptableException {
 		// TODO: for now just check the query parameter. later 'Accept' header.
-		// Allow for application/json or json as values
-		ContentType contentTypeLocal = null;
+		ContentType contentType = null;
 		if (ContentType.json.name().equalsIgnoreCase(mimeType)) {
-			contentTypeLocal = ContentType.json;
+			contentType = ContentType.json;
 		} else if (ContentType.waterml.name().equalsIgnoreCase(mimeType)) {
-			contentTypeLocal = ContentType.waterml;
+			contentType = ContentType.waterml;
 		}
 
-		if (contentTypeLocal == null) {
+		if (contentType == null) {
 			throw new HttpMediaTypeNotAcceptableException(CONTENT_TYPE_MESS + acceptedTypes);
-		} else if(acceptedTypes != null && !acceptedTypes.contains(contentTypeLocal)) {
+		} else if(acceptedTypes != null && !acceptedTypes.contains(contentType)) {
 			throw new HttpMediaTypeNotAcceptableException(CONTENT_TYPE_MESS + acceptedTypes);
 		}
-		this.contentType = contentTypeLocal;
-	}
-
-	protected boolean contentIsJson() {
-		return ContentType.json.equals(contentType);
-	}
-
-	protected boolean contentIsWaterML() {
-		return ContentType.waterml.equals(contentType);
+		return contentType;
 	}
 
 }
