+++ src/main/java/gov/usgs/wma/waterdata/collections/CollectionsController.java
@@ -1,5 +1,6 @@
 package gov.usgs.wma.waterdata.collections;
 
+import gov.usgs.wma.waterdata.BaseController;
 import gov.usgs.wma.waterdata.OgcException;
 import gov.usgs.wma.waterdata.openapi.schema.collections.CollectionGeoJSON;
 import gov.usgs.wma.waterdata.openapi.schema.collections.CollectionsGeoJSON;
