+++ src/main/java/gov/usgs/wma/waterdata/collections/ObservationsController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import gov.usgs.wma.waterdata.BaseController;
 import gov.usgs.wma.waterdata.OgcException;
 import gov.usgs.wma.waterdata.openapi.schema.observations.ObservationTypesJSON;
 import gov.usgs.wma.waterdata.openapi.schema.observations.ObservationsJSON;
