+++ src/main/java/gov/usgs/wma/waterdata/parameter/ContentType.java
@@ -3,4 +3,12 @@ package gov.usgs.wma.waterdata.parameter;
 public enum ContentType {
 	json,
 	waterml;
+
+	public boolean isJson() {
+		return this.equals(json);
+	}
+
+	public boolean isWaterML() {
+		return this.equals(waterml);
+	}
 }
